SeNorge -> MtClim input
Convert the SeNorge binary data to the gridded based text files. Here I write the output once per month to avoid excessive time of opening files for each day and to ensure not to run out of the memory. Still need to think how to write the output into the binary files. The "writeBin" function seems not have the "append" option as the "write.table".